Fix most warnings related to bad arguments for CPLError calls, revealed by previous commit
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@14048 f0d54148-0727-0410-94bb-9a71ac55c965

+++ port/cpl_minixml.cpp
@@ -1719,7 +1719,7 @@ CPLXMLNode *CPLParseXMLFile( const char *pszFilename )
         CPLError( CE_Failure, CPLE_OutOfMemory, 
                   "Out of memory allocating space for %d byte buffer in\n"
                   "CPLParseXMLFile(%.500s).", 
-                  nLen+1, pszFilename );
+                  (int)nLen+1, pszFilename );
         VSIFCloseL( fp );
         return NULL;
     }
@@ -1727,7 +1727,7 @@ CPLXMLNode *CPLParseXMLFile( const char *pszFilename )
     {
         CPLError( CE_Failure, CPLE_FileIO, 
                   "VSIFRead() result short of expected %d bytes from %.500s.", 
-                  nLen, pszFilename );
+                  (int)nLen, pszFilename );
         pszDoc[0] = '\0';
     }
     VSIFCloseL( fp );
